nondocker emits exception on failure
