+++ airbyte-cdk/bulk/core/base/src/main/kotlin/io/airbyte/cdk/ConnectorUncleanExitException.kt
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
+ */
+
+package io.airbyte.cdk
+
+class ConnectorUncleanExitException(val exitCode: Int) :
+    Exception("Destination process exited uncleanly: $exitCode")
